tech-exercise/web-app/gensen/requirements.txt
anthropic==0.49.0
flask==3.0.2
mysql-connector-python==8.3.0
rich==13.7.0